Move first `MerlinArray` design option into `convertible` directory
