+++ merlin/array/convertible/convertible.py
