+++ tests/unit/array/convertible/test_base.py
